Align current index when the viewer gets the focus
Fixes #21

This is important because in the preview mode the user can alter the
current index (and the selection), then operate in the preview and
either fail (because no image is current/selected) or operate on a
different one (resp. an unexpected? selection of images)

Todo: figure whether to limit this to the case where no image is
selected because setting a different current index will always change
the preview and maybe it's desireable to work on multiple selected
images despite operating from the preview while at the other hand it's
super annoying to lose the selection when clicking the preview