Queue connection that loads images on index changes
The direct connection can result in recursive behavior where
an index gets activated, the image loads, that takes some time
events are processed, the user activates another image, that triggers
the connection and so the second index change connection becomes nested
inside the first.

One consequence is that the selection jumps back to the outer (previous)
image while the current index stays at the inner (second) image, causing
glitchy behavior and at worst the user is confused about the current selection
and deletes the wrong image.

Afaict the QueuedConnection is sufficient to postpone the loading behind the
selection update.

For the same reason the loading has to be the last action, otherwise we end
up w/ the wrong info and window title.