Stall thumbsview during image viewer
When showing the viewer, docks get hidden, the thumbsview
resizes and the reverse process happens when hiding the
viewer. The process doesn't maintain scroll position, so
we scroll to the current index (anyway, since we might have
changed it) but that's not necessarily to the spot wehere
we left it.
So don't resize during this this.

Fixes #35