conditionally guard enter key
The table propagates key events to the window - which is nice
because despite it having the focus, enter accepts the dialog
What is not nice is that the cell editor doesn't always handle
this and the enter event gets propagated when entering a new
file name - resulting in either accidentally accepting the copy
or, if there's a pending qmessagebox, crash because of the second
event loop.
So handle this with an event filter.